refactor(entity-set): move specific code to own file
