+++ __tests__/schema/entity/set.ts
@@ -67,10 +67,8 @@ import { autoreviewTaxonomyIds } from '~/config/autoreview-taxonomies'
 import { Model } from '~/internals/graphql'
 import { DatabaseLayer } from '~/model'
 import { castToUuid, EntityRevisionType, EntityType } from '~/model/decoder'
-import {
-  SetAbstractEntityInput,
-  fromEntityTypeToEntityRevisionType,
-} from '~/schema/uuid/abstract-entity/utils'
+import { SetAbstractEntityInput } from '~/schema/uuid/abstract-entity/entity-set-handler'
+import { fromEntityTypeToEntityRevisionType } from '~/schema/uuid/abstract-entity/utils'
 
 class EntitySetWrapper {
   public mutationName: string
