+++ packages/server/src/schema/uuid/abstract-entity/utils.ts
@@ -19,35 +19,22 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
-import * as serloAuth from '@serlo/authorization'
 import { UserInputError } from 'apollo-server'
 import * as t from 'io-ts'
 import R from 'ramda'
 
-import { getTaxonomyTerm } from '../taxonomy-term/utils'
-import { autoreviewTaxonomyIds } from '~/config/autoreview-taxonomies'
 import { InvalidCurrentValueError } from '~/internals/data-source-helper'
 import {
-  assertArgumentIsNotEmpty,
-  assertUserIsAuthenticated,
-  assertUserIsAuthorized,
   Context,
   Model,
   PickResolvers,
   Repository,
   ResolverFunction,
 } from '~/internals/graphql'
-import {
-  EntityDecoder,
-  EntityRevisionType,
-  EntityType,
-  TaxonomyTermDecoder,
-} from '~/model/decoder'
-import { fetchScopeOfUuid } from '~/schema/authorization/utils'
+import { EntityDecoder, EntityRevisionType, EntityType } from '~/model/decoder'
 import { Connection } from '~/schema/connection/types'
 import { createRepositoryResolvers } from '~/schema/uuid/abstract-repository/utils'
 import { VideoRevisionsArgs } from '~/types'
-import { WithRequired } from '~/utils'
 
 export function createEntityResolvers<
   R extends Model<'AbstractEntityRevision'>
@@ -78,310 +65,7 @@ export function createEntityResolvers<
   }
 }
 
-export interface SetAbstractEntityInput {
-  changes: string
-  subscribeThis: boolean
-  subscribeThisByEmail: boolean
-  needsReview: boolean
-  entityId?: number
-  parentId?: number
-  cohesive?: 'true' | 'false'
-  content?: string
-  description?: string
-  metaDescription?: string
-  metaTitle?: string
-  title?: string
-  url?: string
-}
-
-interface setEntityMutationArgs {
-  entityType: EntityType
-  input: SetAbstractEntityInput
-  mandatoryFields: { [key: string]: string | boolean }
-}
-
-export async function buildSetEntityResolver<
-  C extends Model<'AbstractEntity'> & {
-    taxonomyTermIds?: number[]
-    parentId?: number
-  },
-  P extends Model<'AbstractEntity'> & {
-    taxonomyTermIds?: number[]
-    parentId?: number
-  }
->(
-  args: setEntityMutationArgs,
-  context: Context,
-  {
-    childDecoder,
-    parentDecoder,
-  }: { childDecoder: t.Type<C, unknown>; parentDecoder?: t.Type<P, unknown> }
-) {
-  const { userId, dataSources } = context
-
-  assertUserIsAuthenticated(userId)
-
-  const { entityType, input, mandatoryFields } = args
-
-  assertArgumentIsNotEmpty(mandatoryFields)
-
-  const { entityId, parentId } = input
-
-  if (userWantsToCreateNewEntity(entityId, parentId)) {
-    const parentTypename = (
-      await dataSources.model.serlo.getUuid({ id: parentId })
-    )?.__typename
-
-    const newInput = {
-      ...input,
-      taxonomyTermId: parentTypename === 'TaxonomyTerm' ? parentId : undefined,
-      parentId: parentTypename === 'TaxonomyTerm' ? undefined : parentId,
-    }
-    return await buildCreateEntityResolver(
-      {
-        entityType,
-        input: newInput,
-      },
-      { userId, dataSources }
-    )
-  } else {
-    return await buildAddRevisionResolver(
-      {
-        revisionType: fromEntityTypeToEntityRevisionType(entityType),
-        input: {
-          ...input,
-          // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-          entityId: entityId!,
-        },
-      },
-      { userId, dataSources },
-      { childDecoder, parentDecoder }
-    )
-  }
-}
-
-function userWantsToCreateNewEntity(
-  entityId?: number,
-  parentId?: number
-): parentId is number {
-  if ((!entityId && !parentId) || (entityId && parentId))
-    throw new UserInputError('Either entityId or parentId has to be provided')
-
-  if (entityId) {
-    return false
-  }
-  return true
-}
-
-async function buildCreateEntityResolver(
-  {
-    entityType,
-    input,
-  }: {
-    entityType: EntityType
-    input: Omit<SetAbstractEntityInput, 'entityId'> & {
-      taxonomyTermId?: number
-    }
-  },
-  {
-    dataSources,
-    userId,
-  }: { dataSources: Context['dataSources']; userId: number }
-) {
-  const {
-    changes,
-    needsReview,
-    subscribeThis,
-    subscribeThisByEmail,
-    parentId,
-    taxonomyTermId,
-    ...inputFields
-  } = input
-
-  let parent: Model<'AbstractEntity' | 'TaxonomyTerm'>
-
-  if (taxonomyTermId) {
-    parent = await getTaxonomyTerm(taxonomyTermId, dataSources)
-  } else if (parentId) {
-    parent = await getEntity(parentId, dataSources)
-  } else {
-    throw new Error('parentId or taxonomyTermId has to be provided')
-  }
-
-  await assertUserIsAuthorized({
-    userId,
-    dataSources,
-    message: 'You are not allowed to add revision to this entity.',
-    guard: serloAuth.Uuid.create('Entity')(
-      serloAuth.instanceToScope(parent.instance)
-    ),
-  })
-
-  const fields = removeUndefinedFields(
-    inputFields as { [key: string]: string | undefined }
-  )
-
-  const entity = await dataSources.model.serlo.createEntity({
-    entityType,
-    userId,
-    input: {
-      changes,
-      instance: parent.instance,
-      licenseId: 1,
-      needsReview,
-      parentId,
-      subscribeThis,
-      subscribeThisByEmail,
-      taxonomyTermId,
-      fields,
-    },
-  })
-
-  return {
-    record: entity,
-    success: entity != null,
-    query: {},
-  }
-}
-
-async function buildAddRevisionResolver<
-  C extends Model<'AbstractEntity'> & {
-    taxonomyTermIds?: number[]
-    parentId?: number
-  },
-  P extends Model<'AbstractEntity'> & {
-    taxonomyTermIds?: number[]
-    parentId?: number
-  }
->(
-  {
-    input,
-    revisionType,
-  }: {
-    revisionType: EntityRevisionType
-    input: WithRequired<Omit<SetAbstractEntityInput, 'parentId'>, 'entityId'>
-  },
-  {
-    dataSources,
-    userId,
-  }: { dataSources: Context['dataSources']; userId: number },
-  {
-    childDecoder,
-    parentDecoder,
-  }: { childDecoder: t.Type<C, unknown>; parentDecoder?: t.Type<P, unknown> }
-) {
-  const {
-    entityId,
-    changes,
-    needsReview,
-    subscribeThis,
-    subscribeThisByEmail,
-    ...inputFields
-  } = input
-
-  const scope = await fetchScopeOfUuid({
-    id: entityId,
-    dataSources,
-  })
-
-  await assertUserIsAuthorized({
-    userId,
-    dataSources,
-    message: 'You are not allowed to add revision to this entity.',
-    guard: serloAuth.Uuid.create('EntityRevision')(scope),
-  })
-
-  const entity = await dataSources.model.serlo.getUuidWithCustomDecoder({
-    id: entityId,
-    decoder: childDecoder,
-  })
-
-  let isAutoreviewEntity = false
-
-  if (entity.taxonomyTermIds) {
-    isAutoreviewEntity = await verifyAutoreviewEntity(
-      entity.taxonomyTermIds,
-      dataSources
-    )
-  }
-
-  if (entity.parentId && parentDecoder) {
-    const parent = await dataSources.model.serlo.getUuidWithCustomDecoder({
-      id: entity.parentId,
-      decoder: parentDecoder,
-    })
-
-    if (parent.taxonomyTermIds) {
-      isAutoreviewEntity = await verifyAutoreviewEntity(
-        parent.taxonomyTermIds,
-        dataSources
-      )
-    }
-  }
-
-  if (!isAutoreviewEntity && !needsReview) {
-    await assertUserIsAuthorized({
-      userId,
-      dataSources,
-      message: 'You are not allowed to skip the reviewing process.',
-      guard: serloAuth.Entity.checkoutRevision(scope),
-    })
-  }
-
-  const fields = removeUndefinedFields(
-    inputFields as { [key: string]: string | undefined }
-  )
-
-  const { success, revisionId } =
-    await dataSources.model.serlo.addEntityRevision({
-      revisionType,
-      userId,
-      input: {
-        changes,
-        entityId,
-        needsReview: isAutoreviewEntity ? false : needsReview,
-        subscribeThis,
-        subscribeThisByEmail,
-        fields,
-      },
-    })
-
-  return {
-    revisionId,
-    success,
-    query: {},
-  }
-}
-
-export async function verifyAutoreviewEntity(
-  taxonomyTermIds: number[],
-  dataSources: Context['dataSources']
-): Promise<boolean> {
-  return (
-    await Promise.all(
-      taxonomyTermIds.map((id) => checkAnyParentAutoreview(id, dataSources))
-    )
-  ).every((x) => x)
-}
-
-async function checkAnyParentAutoreview(
-  taxonomyTermId: number,
-  dataSources: Context['dataSources']
-): Promise<boolean> {
-  if (autoreviewTaxonomyIds.includes(taxonomyTermId)) return true
-
-  const taxonomyTerm = await dataSources.model.serlo.getUuidWithCustomDecoder({
-    id: taxonomyTermId,
-    decoder: TaxonomyTermDecoder,
-  })
-
-  return (
-    taxonomyTerm.parentId !== null &&
-    (await checkAnyParentAutoreview(taxonomyTerm.parentId, dataSources))
-  )
-}
-
-function removeUndefinedFields(inputFields: {
+export function removeUndefinedFields(inputFields: {
   [key: string]: string | undefined
 }) {
   return R.filter((value) => value != undefined, inputFields) as {
@@ -389,7 +73,7 @@ function removeUndefinedFields(inputFields: {
   }
 }
 
-async function getEntity(
+export async function getEntity(
   entityId: number,
   dataSources: Context['dataSources']
 ) {
@@ -401,7 +85,10 @@ async function getEntity(
   })
 }
 
-async function assertIsEntity(id: number, dataSources: Context['dataSources']) {
+export async function assertIsEntity(
+  id: number,
+  dataSources: Context['dataSources']
+) {
   try {
     await dataSources.model.serlo.getUuidWithCustomDecoder({
       id,
